backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
google-generativeai==0.8.3
elevenlabs==1.9.0
snowflake-connector-python==3.12.4
pydantic==2.9.2
httpx==0.27.0
